Replace old file and update readme
